Turned user widget into button
