+++ Assets/GameCode/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0b8ffb3de66e132438b913d3f5c8947f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
